move shareButton to dedicated component-folder
