+++ src/modules/toolbox/components/shareButton/ShareButton.js
