+++ src/modules/toolbox/index.js
@@ -1,6 +1,7 @@
 import './components/toolContainer';
 import './components/drawToolContent';
 import './components/measureToolContent';
+import './components/shareButton';
 import './components/shareToolContent';
 import './components/toolBar';
 import './i18n';
